+++ ebl/corpus/infrastructure/mongo_text_repository_query.py
@@ -1,4 +1,4 @@
-from typing import List, Optional, Tuple, Sequence, Dict, TypedDict, cast
+from typing import List, Optional, Tuple, Sequence, Dict
 from pymongo.collation import Collation
 
 from ebl.common.query.query_result import CorpusQueryResult
@@ -7,16 +7,10 @@ from ebl.corpus.application.schemas import (
     ChapterSchema,
     DictionaryLineSchema,
     ManuscriptSchema,
-    ManuscriptAttestationSchema,
-    UncertainFragmentAttestationSchema,
 )
 from ebl.corpus.domain.chapter import Chapter, ChapterId
 from ebl.corpus.domain.dictionary_line import DictionaryLine
 from ebl.corpus.domain.manuscript import Manuscript
-from ebl.corpus.domain.manuscript_attestation import ManuscriptAttestation
-from ebl.corpus.domain.uncertain_fragment_attestation import (
-    UncertainFragmentAttestation,
-)
 from ebl.corpus.infrastructure.chapter_query_filters import (
     filter_query_by_transliteration,
 )
@@ -26,45 +20,21 @@ from ebl.corpus.infrastructure.manuscript_lemma_filter import (
 )
 from ebl.corpus.infrastructure.queries import (
     chapter_id_query,
-    join_text,
     join_text_title,
 )
 from ebl.errors import NotFoundError
 from ebl.fragmentarium.infrastructure.queries import is_in_fragmentarium, join_joins
-from ebl.transliteration.application.museum_number_schema import MuseumNumberSchema
 from ebl.transliteration.domain.genre import Genre
-from ebl.transliteration.domain.museum_number import MuseumNumber
 from ebl.transliteration.domain.transliteration_query import TransliterationQuery
 from ebl.corpus.infrastructure.mongo_text_repository_base import (
-    MongoTextRepositoryBase,
     chapter_not_found,
 )
-from ebl.corpus.application.text_repository import CorpusFragmentsMapping
+from ebl.corpus.infrastructure.mongo_text_repository_query_fragment import (
+    MongoTextRepositoryQueryFragment,
+)
 
 
-class MuseumNumberMapping(TypedDict):
-    prefix: str
-    number: str
-    suffix: str
-
-
-def get_museum_number_mappings(
-    museum_numbers: Sequence[MuseumNumber],
-) -> List[MuseumNumberMapping]:
-    return [
-        cast(
-            MuseumNumberMapping,
-            {
-                "prefix": museum_number["prefix"],
-                "number": museum_number["number"],
-                "suffix": museum_number["suffix"],
-            },
-        )
-        for museum_number in MuseumNumberSchema().dump(museum_numbers, many=True)
-    ]
-
-
-class MongoTextRepositoryQuery(MongoTextRepositoryBase):
+class MongoTextRepositoryQuery(MongoTextRepositoryQueryFragment):
     def query_by_transliteration(
         self, query: TransliterationQuery, pagination_index: int
     ) -> Tuple[Sequence[Chapter], int]:
@@ -226,93 +196,3 @@ class MongoTextRepositoryQuery(MongoTextRepositoryBase):
             )
         except NotFoundError as error:
             raise chapter_not_found(id_) from error
-
-    def query_corpus_by_manuscripts(
-        self, museum_numbers: List[MuseumNumber]
-    ) -> List[ManuscriptAttestation]:
-        cursor = self._chapters.aggregate(
-            [
-                {"$unwind": "$manuscripts"},
-                {
-                    "$set": {
-                        "museumNumbers": {
-                            "prefix": "$manuscripts.museumNumber.prefix",
-                            "number": "$manuscripts.museumNumber.number",
-                            "suffix": "$manuscripts.museumNumber.suffix",
-                        }
-                    }
-                },
-                {
-                    "$match": {
-                        "museumNumbers": {
-                            "$in": get_museum_number_mappings(museum_numbers)
-                        }
-                    },
-                },
-                {
-                    "$replaceRoot": {
-                        "newRoot": {
-                            "chapterId": {
-                                "textId": "$textId",
-                                "stage": "$stage",
-                                "name": "$name",
-                            },
-                            "manuscript": "$manuscripts",
-                        }
-                    }
-                },
-                *join_text(),
-                {"$unwind": "$text"},
-            ]
-        )
-        return ManuscriptAttestationSchema().load(cursor, many=True)
-
-    def query_corpus_by_uncertain_fragments(
-        self, museum_numbers: List[MuseumNumber]
-    ) -> List[UncertainFragmentAttestation]:
-        cursor = self._chapters.aggregate(
-            [
-                {"$unwind": "$uncertainFragments"},
-                {
-                    "$set": {
-                        "museumNumbers": {
-                            "prefix": "$uncertainFragments.prefix",
-                            "number": "$uncertainFragments.number",
-                            "suffix": "$uncertainFragments.suffix",
-                        }
-                    }
-                },
-                {
-                    "$match": {
-                        "museumNumbers": {
-                            "$in": get_museum_number_mappings(museum_numbers)
-                        }
-                    },
-                },
-                {
-                    "$replaceRoot": {
-                        "newRoot": {
-                            "chapterId": {
-                                "textId": "$textId",
-                                "stage": "$stage",
-                                "name": "$name",
-                            },
-                            "museumNumber": "$uncertainFragments",
-                        }
-                    }
-                },
-                *join_text(),
-                {"$unwind": "$text"},
-            ]
-        )
-        return UncertainFragmentAttestationSchema().load(cursor, many=True)
-
-    def query_corpus_by_related_fragments(
-        self, museum_numbers: List[MuseumNumber]
-    ) -> CorpusFragmentsMapping:
-        return {
-            "manuscripts": self.query_corpus_by_manuscripts(museum_numbers),
-            "uncertain_fragments": self.query_corpus_by_uncertain_fragments(
-                museum_numbers
-            ),
-        }
